Extract alphamodulated imagesource
